format attachments
