+++ includes/Core/Mail/WeMailMailer55.php
@@ -5,10 +5,8 @@ use PHPMailer\PHPMailer\PHPMailer;
 use PHPMailer\PHPMailer\Exception as PHPMailerException;
 
 class WeMailMailer55 extends PHPMailer {
-    /**
-     * @var $phpmailer PHPMailer
-     */
-    protected $phpmailer;
+
+    use MailerHelper;
 
     /**
      * Overwrite phpmailer send method
@@ -17,11 +15,11 @@ class WeMailMailer55 extends PHPMailer {
      */
     public function send() {
         $response = wemail()->api->emails()->transactional()->post( array(
-            'to' => $this->formatEmailAddress($this->phpmailer->getToAddresses()),
+            'to' => $this->formatEmailAddress( $this->phpmailer->getToAddresses() ),
             'subject' => $this->phpmailer->Subject,
             'message' => $this->phpmailer->Body,
             'type' => $this->phpmailer->ContentType,
-            'attachments' => $this->phpmailer->getAttachments()
+            'attachments' => $this->formatAttachments( $this->phpmailer->getAttachments() )
         ) );
 
         if ( is_wp_error( $response ) ) {
@@ -33,26 +31,5 @@ class WeMailMailer55 extends PHPMailer {
         }
 
         return true;
-    }
-
-    /**
-     *  Format Email Addresses
-     *
-     * @param $address
-     * @return array
-     */
-    protected function formatEmailAddress( $address ) {
-        return array_map( function ( $address ) {
-            return $address[0];
-        }, $address );
-    }
-
-    /**
-     * Set Mailer
-     *
-     * @param $mailer
-     */
-    public function setPHPMailer( $mailer ) {
-        $this->phpmailer = $mailer;
     }
 }
